Fix return type of delete queries - hibernate does not support long type

+++ contact/src/main/java/com/cleevio/vexl/module/contact/service/ContactRepository.java
@@ -109,11 +109,11 @@ interface ContactRepository extends JpaRepository<UserContact, Long>, JpaSpecifi
     Set<User> retrieveSecondDegreeFirebaseTokensByHashes(String newUserHash, ConnectionLevel level, Set<String> importedHashes);
 
     @Query("select count(uc) from UserContact uc")
-    long getConnectionsCount();
+    int getConnectionsCount();
 
     @Query(value = "select count(unique uc.hash_from) from UserContact uc", nativeQuery = true)
-    long getCountOfUsers();
+    int getCountOfUsers();
 
     @Query(value = "select count(unique uc.hash_to) from UserContact uc", nativeQuery = true)
-    long getCountOfContacts();
+    int getCountOfContacts();
 }
